+++ apis/api-media/infrastructure/locals.tf
@@ -25,6 +25,7 @@ locals {
     "GOOGLE_APPLICATION_JSON",
     "INTEROP_TOKEN",
     "JOURNEYS_SHORTLINK_DOMAIN",
+    "MEDIA_TRANSCODER_TASK_DEFINITION",
     "MUX_ACCESS_TOKEN_ID",
     "MUX_SECRET_KEY",
     "MUX_UGC_ACCESS_TOKEN_ID",
