+++ apis/api-media/src/schema/cloudflare/r2/index.ts
@@ -1,2 +1,3 @@
 import './asset'
 import './inputs'
+import './transcode'
