+++ apis/api-media/src/schema/video/index.ts
@@ -7,4 +7,3 @@ import './videoSnippet'
 import './videoStudyQuestion'
 import './videoSubtitle'
 import './videoTitle'
-import './videoTranscode'
