+++ apis/api-media/src/workers/assetTranscode/config.ts
@@ -1,0 +1,1 @@
+export const queueName = 'api-media-asset-transcode-return'
