--- apis/api-media/src/workers/videoTranscode/config.ts
@@ -1,2 +1,0 @@
-export const queueName = 'api-media-video-transcode'
-export const jobName = `${queueName}-job`
