--- apis/api-media/src/workers/videoTranscode/index.ts
@@ -1,2 +1,0 @@
-export { service } from './service'
-export { jobName, queueName } from './config'
