--- apis/api-media/src/workers/videoTranscode/service/index.ts
@@ -1,1 +1,0 @@
-export { service } from './service'
